Added tests and perf for unrolled versions
