+++ perf/evd/cyclic/evd_cyclic_perf.cpp
@@ -7,14 +7,15 @@
 #include "types.hpp"
 
 using EVDEpochType = decltype(&evd_cyclic);
+using EVDEpochType = decltype(&evd_cyclic_oneloop);
+using EVDEpochType = decltype(&evd_cyclic_unroll_outer);
+using EVDEpochType = decltype(&evd_cyclic_unroll_inner);
 using EVDTolType = decltype(&evd_cyclic_tol);
 
-std::vector<EVDEpochType> epoch_based_versions = {
-    evd_cyclic,
-};
-std::vector<std::string> epoch_based_names = {
-    "evd_cyclic",
-};
+std::vector<EVDEpochType> epoch_based_versions = {evd_cyclic, evd_cyclic_oneloop, evd_cyclic_unroll_outer,
+                                                  evd_cyclic_unroll_inner};
+std::vector<std::string> epoch_based_names = {"evd_cyclic", "evd_cyclic_oneloop", "evd_cyclic_unroll_outer",
+                                              "evd_cyclic_unroll_inner"};
 
 std::vector<EVDTolType> tol_based_versions = {
     evd_cyclic_tol,
@@ -64,7 +65,7 @@ using CostFuncType = decltype(&base_cost_evd);
 using CostFuncType = decltype(&tol_cost);
 
 std::vector<CostFuncType> tol_based_cost_fns = {tol_cost};
-std::vector<CostFuncType> evdcyclic_epoch_based_cost_fns = {base_cost_evd};
+std::vector<CostFuncType> evdcyclic_epoch_based_cost_fns = {base_cost_evd, one_loop_cost, one_loop_cost, one_loop_cost};
 
 int main() {
     size_t n;
