+++ src/evd/cyclic/CMakeLists.txt
@@ -1,5 +1,6 @@
 set(SOURCES
     evd_cyclic.cpp
+    evd_cyclic_unroll.cpp
     evd_cyclic.hpp
     )
